Initial Commit
Simple script that lets you automatically teleport players to a named destination when they hit a floor. Does not affect surfing. Affects spines. Affects displacements. Can add a targetname 'safe' to areas like bhops and slides that should allow being on the ground.

It works by adding a targetname to the player which matches a info_teleport_destination name, if you are ON GROUND and have this targetname you get teleported.